+++ shotover/src/lib.rs
@@ -66,10 +66,7 @@ pub mod connection;
 mod connection_span;
 pub mod frame;
 pub mod hot_reload;
-pub mod hot_reload_client;
-pub mod hot_reload_server;
 mod http;
-pub mod json_parsing;
 pub mod message;
 mod observability;
 pub mod runner;
