+++ shotover/src/runner.rs
@@ -187,7 +187,7 @@ impl Shotover {
     ) -> Result<()> {
         if let Some(socket_path) = hotreload_from_socket.clone() {
             info!("Hot reload CLIENT mode - requesting socket handoff from existing shotover");
-            crate::hot_reload_client::perform_hot_reloading(socket_path)
+            crate::hot_reload::client::perform_hot_reloading(socket_path)
                 .await
                 .context("Hot reload client failed")?;
         }
@@ -198,7 +198,7 @@ impl Shotover {
 
             let socket_path = hotreload_socket_path.clone();
             tokio::spawn(async move {
-                match crate::hot_reload_server::UnixSocketServer::new(socket_path) {
+                match crate::hot_reload::server::UnixSocketServer::new(socket_path) {
                     Ok(mut server) => {
                         info!("Unix socket server started for hot reload communication");
                         if let Err(e) = server.run().await {
